rev Auth Respository and Auth Service

+++ internal/core/domain/entity/auth_entity.go
@@ -1,11 +1,12 @@
 package entity
 
+
 type LoginRequest struct {
-	Email string 
+	Email string
 	Password string
 }
 
-type AccessToken struct{
-	AccessToken string 
-	ExpiresAt int
+type AccessToken struct {
+	AccessToken string
+	ExpiresAt int64	 	
 }
